common: add STATIC_IF and STATIC_IF_NOT macros
A common pattern to use with IS_ENABLED is like this:

/*
 * This var should only be used if CONFIG_FOO. The linker errors if
 * CONFIG_FOO is not defined is intentional.
 */
 #ifdef CONFIG_FOO
 static
 #else
 extern
 #endif
 int some_var;

The issue with this is that it leads to an over-verbose and
potentially hard to read pattern, and does not have the check that
CONFIG_FOO was only defined to blank.

Suppose a macro like this existed:

STATIC_IF(CONFIG_FOO) int some_var;

... which expands to "static" when CONFIG_FOO is defined to empty,
"extern" when CONFIG_FOO is not defined, and errors when CONFIG_FOO is
defined to non-empty.

This CL implements that, as well as the inverse (STATIC_IF_NOT).

BUG=chromium:989786
BRANCH=none
TEST=provided unit tests, buildall

Change-Id: Ib57aaba62bc184fda9aa782a780d5f13ba44ae88
Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>