hatch: Add charger-related configs for low-battery booting
When there is a battery that isn't communicating, the DUT should not
boot the AP in order to avoid brown outs after the system jumps to RW
and can begin PD negotiation.

BUG=b:136970148
BRANCH=none
TEST=Disable PD negotiation in RO, cutoff battery, plug in A/C, and
verify that the DUT stays in G3, and doesn't attempt to boot the AP,
because Hatch's battery is slow to come back from cutoff (>10 s).

Change-Id: I6f53ce83e89520369e9761a64133839b3c162b91
Signed-off-by: Tim Wawrzynczak <twawrzynczak@chromium.org>