servo_v4: Support DP alt-mode
This is the initial change to support DP alt-mode, with some TODO
items to make some settings more configurable for testing.

DP alt-mode is only enabled when
 * servo v4 is configured dts-mode disabled, like through the
   command 'dut-control servo_v4_dts_mode:off', and
 * a monitor is plugged to the mini-DP port (for minimizing the impact
   to PD FAFT).

The realtime HPD monitoring is the following patch.

There is a bug (b/140900633) in servo v4 hardware that only 2-lane DP
works.

BRANCH=servo
BUG=b:137219603
TEST=Configured to disable dts-mode, plugged the DP cable to monitor,
plugged the Type-C cable to DUT, the external monitor worked.

Change-Id: I129754afe1b13fe349d472de4ee6bcec3fd8b854
Signed-off-by: Wai-Hong Tam <waihong@google.com>