cr50: extend INT_AP_L pulse
This patch increases INT_AP_L so that it can meet the requirement
from some platforms regarding dynamic power gating. It used to be
in a range of 1.9~6 usec, and now it is 10~14 usec.
The pulse is overlapped with some codes processing to avoid any
time extension at CR0 IRQ handler.

BUG=b:130515803
BRANCH=cr50
TEST=checked INT_AP_L pulse length ranges extended with logic
analyzer on Atlas and Sarien.
Checked dmesg and coreboot log has no TPM errors.

Change-Id: Iea8d0a779fff7cbda0c8647f3c1de719c3c3d7e0
Signed-off-by: Namyoon Woo <namyoon@chromium.org>