printf: Convert %b to %pb
In order to turn on compile-time printf format checking,
non-standard specifiers like %b (binary) must be removed. Convert
that into %pb, which takes a pointer to an unsigned integer to
convert into binary, since %p is the standard-approved way to
extend printf.

For now, disallow %lpb, which would print a 64-bit integer in
binary. The intbuf in the printf function is currently 34 bytes
long, and would have to be extended to 66 bytes to support
64-bit binary integers. Disallow %lpb for now so if we want to
support this later we don't need to worry about regressions.

BUG=chromium:984041
TEST=make -j buildall
BRANCH=None

Change-Id: I8ef995dcf97af688fbca98ab6ff59b84092b69e3
Signed-off-by: Evan Green <evgreen@chromium.org>