wip:usbc: update state machine framework
I still need to add some more comments and clean up organization a
little.

- OBJ is renamed to context (CTX) for current and last state
 - State definition now includes its parent (no need for the boiler
plate function that takes in a signal now)
 - The init_state, set_state, and exe_state have been re-written to take
advantages of new state machine definition. I had to add more logic to
handle hierarchical states fully. (I still need to document the behavior
better when a child states handles (by calling set_state) something and
how the parent states will run)
 - Explicitly create the USB states at the bottom of the file with all
of the statics. Don't need to use macros (even though I did suggest them)
  - Uses null when we do_nothing instead of calling into a function
 - Created a "private" enum in the C file that lists all of the states
in the file, that we can use to refer to a state (it is also the
index into the states array for that state).
 - Changed prototype of state function to return void, since we aren't
really using the return value and it cleans up a lot of return 0 that
aren't needed.
 - Add const to int port since we can (and should) for typedef
 - Moves struct definition to implementation file only to keep
implementation details private. We can access data through accessor if
needed.

Change-Id: I482a63e08f7d63022d5102b891a2fac0b0faa46f