atlas: Check DP MF-bit against selected pin cfg
When we are configuring a Type-C port for DisplayPort alternate mode, we
should check to see that the selected pin config supports multi-function
mode or not.  This commit fixes a bug where we were setting the
SuperSpeed muxes based solely upon the Multi-function Preferred bit in
the DPStatus VDO.  Some Type-C video adapters are buggy and set the MF
preferred bit without actually supporting an MF pin configuration.
Therefore, we trust the reported supported pin configurations in the
DiscMode VDO.

BRANCH=none
BUG=chromium:919756,b:138874110

Change-Id: I946e601ed672031bc4606878531c561cf487a914
Signed-off-by: Matthew Blecker <matthewb@chromium.org>