driver: IO expander: nct38xx: add the interrupt support
The IO pins of Nuvoton TCPC NCT38XX chips have the ability to support
the interrupt function. This commit adds the driver to support it.

However, please note that if the system needs to use an IO on NCT38XX
to support the interrupt, the following two consideration should be
taken into account.
1. Interrupt latency:
    Because it requires to access the registers of NCT38XX via I2C
    transaction to know the interrupt event, there is some added latency
    for the interrupt handling. If the interrupt requires short latency,
    we do not recommend to connect such a signal to the NCT38XX.

2. Shared ALERT pin:
    Because the ALERT pin is shared also with the TCPC ALERT, we do not
    recommend to connect any signal that may generate a high rate of
    interrupts so it will not interfere with the normal work of the
    TCPC.

BRANCH=none
BUG=none
TEST=No error for "make buildall"
TEST=Apply this and related CLs, manually test each IO pins; make sure
each pin's interrupt handler is correctly executed.

Change-Id: I72d835557913d87097b2e0d82165e40fe132ca77
Signed-off-by: CHLin <CHLIN56@nuvoton.com>