Update patch set 16
Patch Set 16:

(1 comment)

> Patch Set 14:
> 
> (2 comments)
> 
> Question: once this is in, shall we switch to using the new verisign function in all ECDSA sign cases (outside of factory cases) and abandon the plain sign? why signature verification was made a mandatory part of signing?

Sorry, this comment slipped my attention. The plan is to switch to verisign function and stop using plain sign. The verification step is necessary to make sure that the adversary did not glitch the signature generation. If the signature generation was somehow compromised, the verification step ensures that we catch it and do not output the signature at all.

Patch-set: 16
Reviewer: Gerrit User 1308870 <1308870@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>