Update patch set 3
Patch Set 3:

> Patch Set 2:
> 
> What are your thoughts on having the ADC always enabled, vs enabling and disabling around each read?

Would use more battery when we are not plugged in and we don't need the voltage ADC conversion.  If we decide to leave it enabled, all we have to do it enable it once and leave it on.  The code will perform one extra read to determine it is already enabled but otherwise will leave it in the state it starts with.

Patch-set: 3
Reviewer: Gerrit User 1346479 <1346479@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>