Update patch set 1
Patch Set 1:

For the record, I am running Sam's new stack (pre-refactor) on a GRL C2 compliance apparatus.

Test results show distinct issues that can rapidly located and fixed. I don't think we should let "perfect" be the enemy of "good" in this case.

I do not feel it is wise to attempt a code refactor at this moment, especially when we may be aiming for RO lock. This time would be better invested fixing functionality, then later approaching the refactor when we have a known baseline to compare to.

Please ping me for the prototype VIF file and GRL C2 reports, as they are large and possibly contain PI.

Patch-set: 1