motion_lid: Notify AP when switched to 360-degree-flipped mode
Assuming a device without hall sensor is switched from clamshell mode
all the way to 360-degree-flipped mode, the event sequence is like:

(a) When the device hits 200-degree, it switches to tablet mode.
(b) We notify AP to re-read DPTF profile from ACPI.
(c) When the device hits 300-degree, it switches to 360-degree
flipped mode and updates DPTF profile in EC space.

Since the DPTF profile is switched in (c) while AP re-reads
the DPTF profile in (b), AP wouldn't update DPTF profile for
360-degree-flipped mode.

To fix this issue, let's notify AP when the device is switched to
360-degree-flipped mode.

BUG=b:139168228
BRANCH=none
TEST=manually test on Nautilus/Hatch and confirm trip temperature
is updated in flipped-360-degree mode

Change-Id: I3bc61fb3d7808bf17f7da7ebc606307fbbf026fc
Signed-off-by: Philip Chen <philipchen@google.com>