cr50: Delete non-volatile counter
This counter was only used by the legacy U2F
implementation, which is no longer required.

This change does not free any space in flash, as
the change that removed the legacy U2F implementation
caused this to already be dropped from the binary.

It does however free the flash space reserved for
the counter itself (2KB per partition), which can now
be repurposed for something else.

BRANCH=none
BUG=b:138459918
TEST=make buildall -j

Signed-off-by: Louis Collard <louiscollard@chromium.org>
Change-Id: I18892e1eb0224b96caa531293403b0b02f28a32b