cr50: Delete non-volatile counter API
This counter was only used by the legacy U2F
implementation, which is no longer required.

This change deletes the code for the counter,
but does not update the flash config to make use
of the pages previously occupied by the counter.

Since this code is already unused, and therefore
already dropped from built firmware images, this
change does not have any impact on image size.

A follow up change can alter the flash config
to reclaim and repurpose the 2KB per partition
previously used by the counter.

BRANCH=none
BUG=b:138459918
TEST=make buildall -j

Signed-off-by: Louis Collard <louiscollard@chromium.org>
Change-Id: I18892e1eb0224b96caa531293403b0b02f28a32b