util: enable chargen for USB console
This patch allows chargen to print output to USB instead of UART,
which is chosen by command parameter.
If USB console is not supported, then the parameter will be ignored,
and output shall go to UART port.
The patch increases flash usage by 48 bytes if CONFIG_CMD_CHARGEN
is defined.

BUG=chromium:992607
BRANCH=None
TEST=manually ran on fleex.

for BOARD in {cr50, fleex}
1. Define CONFIG_CMD_CHARGEN in board/cr50/board.h,
   and baseboard/octopus/baseboard.h.
2. Build binaries, and program them.
3. Connect CCD to Octopus Fleex.
4. Open terminal to Cr50 and EC consoles, and run chargen
   (cr50) chargen 1 4
   >              // no output, because they went to UART.
   (cr50) chargen 1 4 usb
   0000
   >
   (ec) chargen 1 4
   0000
   >
   (ec) chargen 1 4 usb         // usb parameter gets ignored.
   0000
   >

Change-Id: I5810421fef56548e0bd667488e853e724f699a31
Signed-off-by: Namyoon Woo <namyoon@chromium.org>