Update patch set 3
Patch Set 3: Code-Review+1

> Patch Set 3:
> 
> I have not looked at the code changes yet, but have just noticed that this particular patch does not have any impact on the flash use size reported by make.
> 
> Isn't this patch supposed to release the 2K page? Or does the page gets released by the previous patch (in which case I just don't understand how this happens, as the config controlling the allocation of the page is removed by this patch).

Technically this CL doesn't release the counter page (CONFIG_FLASH_NVCTR* stuff is still there, and nvmem starts only after it, so nvmem size hasn't increased), just removes all the code that uses it. So, 'data' flash size is not affected.
And that code is already not a part of the image since the previous CL, which removes the old u2f, because of LTO? So, 'code' flash size was already decreased by the prev CL.

So, lgtm, but we need to think how to make that freed up 2K in 'data' flash useful.

Patch-set: 3
Reviewer: Gerrit User 1152319 <1152319@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1