Update patch set 12
Patch Set 12:

Jerry and I discussed on a separate thread, and it turned out that the infinite loop on write was due to me not configuring the pins for I2C in gpio.inc.

I've now tested the I2C slave support, and it works if I get my host implementation to sleep between writing the command and reading the response. If I don't have that sleep, then the host just reads 0xEC over and over for its responses. The MCU should clock stretch between the response being received and the reply being ready, so that no sleeps are needed on the host side. Please implement that.

Patch-set: 12
Reviewer: Gerrit User 1289152 <1289152@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>