cometlake: Minimize delay for high->low rsmrst passthrough
Hatch designs buffer the PG_EC_RSMRST# signal from the
Silego power good logic through the EC and out EC_PCH_RSMRST# to the
SoC RSMRST# pin. For power off transitions, this should be as fast as
possible, in the ns region if possible. However this time is ~1 msec.

To reduce this delay as much as possible this CL introduces a new
config option CONFIG_CHIPSET_X86_RSMRST_ISR_PASS. When this config is
defined (auto defined for cometlake chipset), the rsmrst signal is
checked in the poewr_signal_interrupt for high->low
transitions. Low->high transistion continue to be handled in the power
state machine.

BUG=b:132421681
BRANCH=None
TEST=Shorted PP1050_A_PG to ground to force an abrupt power down and
then measured time via scope between PG_EC_RSMRST and
EC_PCH_RSMRST. The delay is reduced from ~1 msec to 45 uSec.

Change-Id: I266138a2e235ce47f3060f8e1f6f9bc6a75073ae
Signed-off-by: Scott Collyer <scollyer@google.com>