EFS: Call led_critical on verification error only for Chromeboxes
If a battery isn't charged, we'll hang out in S5 until it's charged.
If a battery is disconnected, the system would boot on PD power.
Either way, a Chromebook will proceed in RO. We don't need to indicate
failure (by LED). Softsync will fix the bad RW (using the traditional
path).

For a Chromebox, EC will stay in S5, requesting recovery with a USB-C
adapter, or boot to recovery mode with a barrel jack adapter. We need
to indicate it on the LED.

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>

BUG=chromium:998135
BRANCH=none
TEST=Boot Nami. Verify EFS and Softsync are working.

Change-Id: Idee6a94f34df5baadf7a09a5b232246b168531b6