Update patch set 8
Patch Set 8: Verified+1

I spent some more time monitoring the behavior of the system before sending to CQ and observed a couple of things I did not notice before: sometimes it still was invoking the unwedge sequence twice in a row, and sometimes it would invoke it even if when SDA line was not held low. There must be some issues with the value read from I2CS_READVAL_SDA, reading it does not always return the state of the SDA line.

I ended up following Ashotosh's suggestion - use a dedicated GPIO to read the actual SDA signal level. Also, after more discussions, it is clear that creating the STOP condition is enough to take the FSM from the 'wedged' state.

This updated variant was tested for a few days continuously resetting the device and the 'wedged' condition happening every 30 resets or so. The i2c interface reliably recovers, there is always a single invocation of the 'unwedge' sequence, the sequence never kicks in unless SDA is low.

Patch-set: 8
Reviewer: Gerrit User 1000039 <1000039@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Verified=+1