Update patch set 10
Patch Set 10:

> I was just tricked by this new naming scheme in cl:1758534:
> 
>   #define I2C_ADDR_EEPROM_FLAGS 0x50
> 
> This doesn't look like an address but just flags related to i2c addr (as amstan pointed out). 
> 
> We might need to revise the naming. If removing _flags isn't preferred, we would rather use longer names such as addr_and_flags or add_w_flags than being confused? Or can we separate addr and flags?

The flags suffix is important, because it is signaling to the developer that there potential flag bits OR'ed with the 7b (or 10b) address. Dennis documented that here [0].

With that said, I can see where confusion could come from with I2C_ADDR_EEPROM_FLAGS. I would be okay with I2C_ADDR_EEPROM_WITH_FLAGS; it is getting a little long, but it is very explicit. If you wanted to put up a CL for that, I would support it. If we are going that route, maybe we also make ADDR_WITH_FLAGS as the suffix so, I2C_EEPROM_ADDR_WITH_FLAGS instead?


> I'm still confused what _flags even means (this commit message doesn't really explain it). As far as I can see everything has "flags" in it. So if everything's the same could we just remove the extra "flags" and go back how it was before?

Flags indicate that there is a potential for bit to bet set in the value that determine functionality such as I2C_FLAG_BIG_ENDIAN. See [0].


Also an unintended helpful consequence of having the _FLAGS suffix is with firmware branches. Having the _FLAGS suffix has generated merge conflict when adding 7-bit i2c address to 8-bit firmware branches (where it would have silently cherry-picked without error if _FLAGS suffix wasn't added for the base address definition).


[0] https://chromium.googlesource.com/chromiumos/platform/ec/+/46fe2ab0e39bf5da7165e4d65a688d6afe273613/include/i2c.h#25

Patch-set: 10
Reviewer: Gerrit User 1259838 <1259838@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>