Update patch set 3
Patch Set 3:

Hold on, we might want to still do this (for every chip too, not just npcx).

The problem right now is that any gpio in the system could either be a local gpio (gpio_get_level) or something on the gpio expander (ioex_get_level). It makes no sense to have separate calling conventions depending on where the gpio is

Please reconsider.

Patch-set: 3