util/getversion: allow non-timestamped build
we normally embed a timestamp into the EC image, but this makes it
difficult to determine if a local change affects the resulting EC
image.  this adds an override to keep the image timestamp constant so
we can easily check if a source tree change affects the resulting EC
image.

Change-Id: Ib4d8069201b85d4ca512e6a9e900d8faa578f5ec
Signed-off-by: Caveh Jalali <caveh@chromium.org>