main: Move non hardware things away from main into init hooks
There's a few things in main that run a little too early.

Some of those things might have more dependencies in the future.

EFS depends on Keyboard which will depend on I2C.

I2C happens in the hooks task which is too late.

TODO TODO.. fill in the rest

BUG=b/138384267
TEST=EC still works
TEST=EFS doesn't happen significantly later than it used to (TODO)
TEST=Recovery keys now work with I2C keyboard
BRANCH=master

Change-Id: I081b85524128491544d480e57fe1ea5c7250a054
Signed-off-by: Alexandru M Stan <amstan@chromium.org>