Update patch set 27
Patch Set 27:

(1 comment)

> >> I'm noticing sometimes the interrupt gets stuck asserted:
> > I have tried press a key and release fast, but I can't duplicate such a problem.
>
> I noticed the pullup matters :) https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/1693863/2..3/board/jacuzzi/gpio.inc
>
> It might have been just that! I can't really reproduce it after adding it.
> 
> I'll investigate some more just in case there's something strange, but I'll probably be nothing.

The only way I know how to reproduce this is by disabling the pullup on the INT pin. Please try something similar.

My first reaction with to problem (without the pullup) is "what's the big deal? there's no reason for the signal to go up after the interrupt gets handled because it's now hiz and nothing's pulling it up".

But I have 2 things that scare me:
* The system goes back to life if I manually wake up the keyscan task
* The INT pin is actually driven low by something (probably IT8801). Even if we added a pullup after we got stuck it wouldn't help us get unstuck.

It's possible the lack of pullup changes the timing of the int just enough it reveals this possible race condition.

The issue is that this stuck state is super bad for the user. Causes the keyboard to not work anymore (possibly until an EC reset or something tickling the keyboard task on the EC).

Patch-set: 27
Reviewer: Gerrit User 1113904 <1113904@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>