drivers/tcs3400: fix a couple values in lux table
A couple of the values in the lux atime table needed adjusting
as they were too aggressive and were causing
train-up / saturate / train-down / repeat... pattern cycles.

Allow device to upshift in training when we're >= 2.5x from
saturation (was >= 3x).

Fixed a couple of nits.

BUG=b:139749448
BRANCH=master
TEST=Walked a flapjack through each lux light level, with a test at
the low end of the level and a test at the high end of each level
(i.e. for LUX level 1000, tested at 1100 and 1900 LUX), assure the
tcs3400 doesn't go into saturation while trying to hit the 90% band.

Change-Id: I84c5bf09e072e96d516d91520a176fa35274ce7a
Signed-off-by: Nick Vaccaro <nvaccaro@google.com>