Update patch set 1
Patch Set 1:

> I could probably be wrong, but seems like Type-C didn't specify any current ramping over 1.5A from bc1.2 DCP. 

That's correct; the BC1.2 spec says 1.5A max, and the Type-C spec echos that.  In reality, BC1.2 was a wild west, and amongst power-hungry devices and proprietary solutions, the industry generally came to the conclusion that a higher max ramp is reasonable.  It's not written, but the Type-C spec is designed to handle the case where BC1.2 goes above 1.5A, so it's really just compatibility at that point.

> If we are going forward with ramp route, what are these empirical voltage/current limitations?

I don't know.  You could look in the code for Spring or Samus or see what other platforms since do.

> According to mt6370 vendor explanation and datasheet, IAICR (0x13[7:2]) is a hard upper bound of input current where the adc reading has +/- 7% tolerance. When IAICR is set to be 2.0A, the typical input current is in fact ~1.86A - causing a mismatch to the current number in discussion.

That's a useful design, since for most cases on Type-C you want to be able to specify a hard maximum current.  BC1.2 is a ramping algorithm, so other than the empirically-determined max current number for compatibility, the mismatch between the register and actual current shouldn't matter.

Patch-set: 1
Reviewer: Gerrit User 1002022 <1002022@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>