ec: create initial EC image for a new variant
Creating an initial EC image for a variant is just a copy of the
base board's files. After all of the CLs to create the variant
have landed, then we can make changes specific to the variant.

BUG=b:140261109
BRANCH=none
TEST=``./initial_ec_image.sh hatch sushi && git show``
Compare board/hatch/* to board/sushi/*

Also run the script with an existing board name to verify that you
can't create a variant that already exists.

Change-Id: Ic54b0b5a8b41476779dcbb6af63bd0ddfaff2896
Signed-off-by: Paul Fagerburg <pfagerburg@chromium.org>