Update patch set 4
Patch Set 4:

why would we want to send chargen to USB console?

The idea is that the source of the chargen stream can not be flow controlled (as is the case when the stream is generated by EC and AP externally to H1) and thus the characters could be dropped.

In case of USB console the characters are generated by the Cr50, there is an inherent flow control of the USB buffer not accepting new characters, the console driver could be waiting on tx_ready(). I would rather not add this to the image as it brings little value as far as testing is concerned.

Patch-set: 4
Reviewer: Gerrit User 1000039 <1000039@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>