fpsensor: Add support for migration to positive match secret.
On receiving an template, if the template has format version 3 and
has empty positive match salt, then generate positive match salt for
this template so that it has positive match secret on the next match.

BRANCH=nocturne
BUG=chromium:927095
TEST=make -j buildall
TEST=tested enrollment, matching, deletion, multifinger on DUT nocturne
TEST=tested migration on DUT nocturne, specifically, enrolled template
with old FPMCU firmware, then updated firmware, verified that templates
are migrated and continue to work.

Change-Id: If8020acfb60a03269af9ddb41023796920116fca
Signed-off-by: Yicheng Li <yichengli@chromium.org>