pd: Extend READY_HOLD_OFF_US timer value for SRC
The READY_HOLD_OFF_US was set to 200 msec for both SNK and SRC
connections. However, in the case of SRC, there can be a race
condition with charge through hubs that initiate a power swap request
and the SRC port sending the VDM discover idenity message.

This CL makes the holdoff timer delay longer for SRC connections and
prefixes READY_HOLD_OFF_US with SNK/SRC to differentiate the two timer
values.

BUG=b:140092163
BRANCH=None
TEST=Tested with the AA-AM1N95W charge through hub. Verfied
that this can be attached to Kohaku with external power already
connected and successfully power role swap into source to charge the
kohaku port. Tested 20-25 connections with no failures. Prior to this
change the failure rate was 1 out of 3 attempts.

Change-Id: I375415fec96baf44ecd4592033e8b44ed4616011
Signed-off-by: Scott Collyer <scollyer@google.com>