npcx: Ensure the watchdog isn't stopped while it is invalid to do so
We cannot unlock the watchdog timer within 3 watch dog ticks of touching
it per the datasheet. This is actually around 100ms so we should protect
against this.

Note: To avoid bringing in other dependencies into the npcx_monitor_fw,
the "old" method of stopping the watchdog is used (no waiting first).

BRANCH=none
BUG=b:140207603
TEST=eliminates cold reset issue.

Change-Id: I440d36fb9aecdd6d78bae6fc4002208198a3357f
Signed-off-by: Jett Rink <jettrink@chromium.org>
Signed-off-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/1842592
Reviewed-by: ML Chao <mlchao@nuvoton.corp-partner.google.com>