Makefile.rules: use patsubst instead of subst
the patsubst function is a better function for removing a path prefix
as it is anchored at the beginning of the search space.  the original
"subst" function can match everywhere in the path which leads to weird
transformations.  patsubst is anchored at the beginning of the string
which better suits the intent here.

BRANCH=none
BUG=none
TEST=make buildall passes, $(out) now only removed at beginning of
	output paths.

Change-Id: Ic5fe7aeb5eb8c87aa8bb49ea0865dc07327cb3e5
Signed-off-by: Caveh Jalali <caveh@chromium.org>