Sweetberry: Fix for struct declaration in power logging.
The struct declaration in Sweetberry was changed while supporting
the new 7-bit address declarations without updating the interface.
This reverses the change on the firmware side so the structs are
correctly unpacked. This structure is only used in the usb_power file
to load the value into a field used internally and during validation
of correctly formated packets.

BUG=b:140969461
BRANCH=None

TEST=Manual testing with powerlog using:
powerlog -b atlas_rev3.board -c atlas_rev3.scenario
Before: No INA output is returned to the console, command fails validation.
After: INA out is returned and visible on console

Signed-off-by: Brian Nemec <bnemec@chromium.org>
Change-Id: Ibddf4b8f8d48af0eea9e219ab3728a95e2cf66a8