Zero data toggles on endpointis when appropriate.
The data toggle should be reset on an endpoint in two circumstances
when the device is configured (with a SET_CONFIGURATION), or whent
the endpoint is "cleared" (with a CLEAR_FEATURE(endpoint_halt).

The endpoint reset code is deficient, in that the data toggle is not
reset when appropriate. The upshot of this is that if the device and
host ever get out of sync, the transfer will be ACKed, but the data
thrown away (which is correct behaviour per spec).

This causes all sorts of head scratching. In this particular case,
the data which is discarded is a mailbox command to a Haven (root
of trust) chip. Thus the host hangs forever waiting for a response
that never comes, because the device never got the command.

This behaviour is seen in testing for reasons which are not clear,
but can be provoked by sending resets to the device, while sending
resets to the device. The device will hang as the following
SET_CONFIGURATION is supposed to clear the toggles (but doesn't)
After implementing this fix, the device no longer hangs after a
reset in this test.

This issue is doubly bad for the device, as the recovery should be
to time out the bulk tranfer, and reset the endpoint, or if that
doesn't work, reset the device. With the deficient code it is
impossible to use standard recovery mechanisms when running across
this issue.

Note: The Synopsis data book notes under the register
(Device Endpoint-n Control Register: DIEPCTLn/DOEPCTLn),
DPID "The application must program the PID of the first packet to
be received or transmitted on this endpoint, after the endpoint is
activated." This is what this change is fixing, the code does not
currently do this.

TEST= Problem can be reproduced by sending a reset to the haven, while
it is being access by USB EC commands. 50% of the time, the next
transfer will hang as the toggle is now incorrect. With this fix, the
device can be reset and no hang.

TEST= A similar test is done by sending CLEAR_FEATURE(endpoint_halt)
commands to the device while processing USB EC commands. On a USB
analyser you can see the toggle has been reset, as the device
continues to work, even if the PIDs before and after the CLEAR_FEATURE
were both DATA0.

BRANCH=NONE

Signed-off-by: barryt@google.com

Change-Id: Id621cfaa422d3589187db7be8a188ed411bb4c98