Zero data toggles on endpointis when appropriate.
The data toggle should be reset on an endpoint in two circumstances
when the device is configured (with a SET_CONFIGURATION), or whent
the endpoint is "cleared" (with a CLEAR_FEATURE(endpoint_halt).

The endpoint reset code is deficient, in that the data toggle is not
reset when appropriate. The upshot of this is that if the device and
host ever get out of sync, the transfer will be ACKed, but the data
thrown away (which is correct behaviour per spec).

This causes all sorts of head scratching. In this particular case,
the data which is discarded is a mailbox command to a Haven (root
of trust) chip. Thus the host hangs forever waiting for a response
that never comes, because the device never got the command.

This behaviour is seen in testing for reasons which are not clear,
but can be provoked by sending resets to the device, while sending
resets to the device. The device will hang as the following
SET_CONFIGURATION is supposed to clear the toggles (but doesn't)
After implementing this fix, the device no longer hangs after a
reset in this test.

This issue is doubly bad for the device, as the recovery should be
to time out the bulk transfer, and reset the endpoint, or if that
doesn't work, reset the device. With the deficient code it is
impossible to use standard recovery mechanisms when running across
this issue.

Note: The Synopsis data book notes under the register
(Device Endpoint-n Control Register: DIEPCTLn/DOEPCTLn),
DPID "The application must program the PID of the first packet to
be received or transmitted on this endpoint, after the endpoint is
activated." This is what this change is fixing, the code does not
currently do this.

BUG=b:141140341
BRANCH=cr50, cr51, cr52
TEST=(cr51) Problem can be reproduced by sending a reset to the
haven, while it is being access by USB EC commands. 50% of the time,
the next transfer will hang as the toggle is now incorrect. With this
fix, the device can be reset and no hang.

TEST=(cr51) A similar test is done by sending
CLEAR_FEATURE(endpoint_halt) commands to the device while processing
USB EC commands. On a USB analyser you can see the toggle has been
reset, as the device continues to work, even if the PIDs before and
after the CLEAR_FEATURE were both DATA0.

TEST=(cr50) Ran flash_ec on fleex through CCD.
Ran uart_stress_test on fleex as well.

Signed-off-by: barryt@google.com

Change-Id: Id621cfaa422d3589187db7be8a188ed411bb4c98
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/1804065
Tested-by: Namyoon Woo <namyoon@chromium.org>
Reviewed-by: Namyoon Woo <namyoon@chromium.org>
Commit-Queue: Namyoon Woo <namyoon@chromium.org>
(cherry picked from commit f6f5fd8659340f6793b723d564f618fb4026e793)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/1920388
Tested-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>
Commit-Queue: Vadim Bendebury <vbendeb@chromium.org>
(cherry picked from commit e56479916264e0a4925f3cd696dc44e8118d4763)