Update patch set 1
Patch Set 1:

The endpoint reset code is deficient, in that the data toggle is not reset when it's invoked. The endpoint is reset on a USB reset, or on a clear endpoint stall. The upshot of this is that if the device and host ever get out of sync, the transfer will be ACKed, but the data thrown away (which is correct behaviour per spec).

This causes all sorts of head scratching. In the particular case I'm looking at, the data which is discarded is a mailbox command to a Haven (root of trust) chip. Thus the host hangs forever waiting for a response that never comes, because the device never got the command.

This behaviour is seen in testing for reasons which are not clear, but can be provoked by sending resets to the device, while sending commands. On 50% of resets, the device will hang on the next reset. After implementing this fix, the device no longer hangs after a reset in this test.

This is doubly bad for the device, as the recovery should be to time out the bulk tranfer, and reset the endpoint, or if that doesn't work, reset the device. As the reset of endpoint or device does not actually do what the spec says it should, it is impossible to use standard recovery mechanisms when running across this issue.

Note: The Synopsis data book notes under the register (Device Endpoint-n Control Register: DIEPCTLn/DOEPCTLn), DPID "The application must program the PID of the first packet to be received or transmitted on this endpoint, after the endpoint is activated." This is what this change is fixing, the code does not currently do this.

Patch-set: 1
Reviewer: Gerrit User 1315467 <1315467@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>