usb_pd: Send SOP' DiscIdent messages on SOP NAKs
Currently, the only scenario where an SOP' message is sent is when we
receive a SOP DiscIdentity Response from the UFP. However, we should be
able to query the cable characteristics even if the UFP responds with a
NAK. So, add handling to send a SOP' DiscIdentity even if we receive a
NAK. This will not otherwise disrupt the PD connection flow.

Note that there may be other situations where SOP' messages can be sent,
but aren't currently being sent.

BUG=chromium:1005941
BRANCH=None
TEST=make -j buildall; Test on a hatch with a Pixel 2 DUT. Cable
     correctly responds with Cable information even after the EC
     receives a UFP DiscIdentity NAK.

Change-Id: Ia856f2eb1547f09ae1dc0318021ad9f2458fc25a
Signed-off-by: Prashant Malani <pmalani@chromium.org>