Update patch set 14
Patch Set 14:

This CL has already been merged, but I would like to understand if the following clause is taken care of (it allows for SOP' communication to be sent without an Explicit Contract):

PD Rev 2.0, Section 2.5.4
"When no Contract or an Implicit Contract is in place (e.g. after a Power Role Swap or Fast Role Swap) only the Source
port that is supplying VCONN is allowed to send packets to a Cable Plug (SOP'/SOP'') and is allowed to respond to
packets from the Cable Plug (SOP'/SOP'') with a GoodCRC in order to discover the Cable Plug’s characteristics (see
Figure 2-2). During this phase all communication with the Cable Plug is initiated and controlled by the Source which
acts to prevent conflicts between SOP* Packets. The Sink does not communicate with the Cable Plug, even if it is the
DFP, and Discards any SOP’ Packets received."

It sounds like the current CL only allows for SOP' DiscIdentity commands to be send when there is an explicit contract.

Patch-set: 14
Reviewer: Gerrit User 1181774 <1181774@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>