Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> > Patch Set 5:
> > 
> > > Patch Set 4:
> > > 
> > > Hi Vijay and Ayushee,
> > > 
> > > Do you test e-mark cable without this patch and the Vconn SRC still can receive cable's ack of discover id request ?
> > 
> > Yes. I can see the cable responding to Discover Identity SOP' even without this patch.
> 
> Hi Ayushee, Could you kindly elaborate on your test set up, and how you trigger the Discover Identity SOP' message from the EC? Thanks!

A e-marked cable is attached between a type-C dock and DUT. Once an explicit contract is established, Discover ID SOP request is sent. On receiving a response, discover ID SOP' is sent and the response from the cable(if any) is stored and the pd flow continues. More can be found in CL:1707851.

Patch-set: 5
CC: Gerrit User 1329037 <1329037@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>